use html inliner and save rendered outputs to the destination directory.

+++ tests/template.rs
@@ -7,9 +7,12 @@ use std::fs;
 fn generate_all_templates() {
     let src_dir = helpers::get_test_dir(&["examples", "src"]);
     let dst_dir = helpers::get_test_dir(&["examples", "dst-generate_all_templates"]);
-    if let Err(_) = fs::remove_dir_all(&dst_dir) {
-        assert!(false); // You need to remove the existing directory before testing it carefully.
+    if dst_dir.exists() {
+        if let Err(_) = fs::remove_dir_all(&dst_dir) {
+            assert!(false); // You need to remove the existing directory before testing it carefully.
+        }
     }
+
     let result = seg::template::generate_all_templates(src_dir, dst_dir);
     assert!(result.is_ok());
 }
